Move type specific methods to hashtable.h. and make them static inline.
Move type specific methods hashtable_add() and hashtable_find() from
hashtable.c to hashtable.h and make them static inline. Update usage
docs to import hashtable.h for type-specific methods. Update sumset.c
and hashtable_test.c for new usage.
