+++ src/sumset.c
@@ -79,7 +79,7 @@ static inline int rs_block_match_cmp(rs_block_match_t *match, const rs_block_sig
 /* Instantiate hashtable for rs_block_sig and rs_block_match. */
 #define ENTRY rs_block_sig
 #define MATCH rs_block_match
-#include "hashtable.c"
+#include "hashtable.h"
 
 /* Get the size of a packed rs_block_sig_t. */
 static inline size_t rs_block_sig_size(const rs_signature_t *sig)
